backend/requirements.txt
flask
flask-cors
requests
vaderSentiment
python-dotenv